Fixed compiler warnings

+++ core/lib/print-stats.c
@@ -49,9 +49,9 @@
 void
 print_stats(void)
 {
-  printf("S %d.%d clock %u tx %lu rx %lu rtx %lu rrx %lu rexmit %lu acktx %lu noacktx %lu ackrx %lu timedout %lu badackrx %lu toolong %lu tooshort %lu badsynch %lu badcrc %lu contentiondrop %lu sendingdrop %lu lltx %lu llrx %lu\n",
+  printf("S %d.%d clock %lu tx %lu rx %lu rtx %lu rrx %lu rexmit %lu acktx %lu noacktx %lu ackrx %lu timedout %lu badackrx %lu toolong %lu tooshort %lu badsynch %lu badcrc %lu contentiondrop %lu sendingdrop %lu lltx %lu llrx %lu\n",
 	 rimeaddr_node_addr.u8[0], rimeaddr_node_addr.u8[1],
-	 clock_time() / CLOCK_SECOND,
+	 (unsigned long)clock_time() / CLOCK_SECOND,
 	 rimestats.tx, rimestats.rx,
 	 rimestats.reliabletx, rimestats.reliablerx,
 	 rimestats.rexmit, rimestats.acktx, rimestats.noacktx,
